Move user data access method from webapps to managers
