+++ test/unit/data/data_access/test_user.py
@@ -1,5 +1,4 @@
 from galaxy.managers import users as lib
-from galaxy.webapps.galaxy.services.users import get_users_for_index
 from . import verify_items
 
 
@@ -36,25 +35,25 @@ def test_get_users_for_index(session, make_user):
     u5 = make_user(email="i", username="z")
     u6 = make_user(email="z", username="i")
 
-    users = get_users_for_index(session, False, f_email="a", expose_user_email=True)
+    users = lib.get_users_for_index(session, False, f_email="a", expose_user_email=True)
     verify_items(users, 1, [u1])
-    users = get_users_for_index(session, False, f_email="c", is_admin=True)
+    users = lib.get_users_for_index(session, False, f_email="c", is_admin=True)
     verify_items(users, 1, [u2])
-    users = get_users_for_index(session, False, f_name="f", expose_user_name=True)
+    users = lib.get_users_for_index(session, False, f_name="f", expose_user_name=True)
     verify_items(users, 1, [u3])
-    users = get_users_for_index(session, False, f_name="h", is_admin=True)
+    users = lib.get_users_for_index(session, False, f_name="h", is_admin=True)
     verify_items(users, 1, [u4])
-    users = get_users_for_index(session, False, f_any="i", is_admin=True)
+    users = lib.get_users_for_index(session, False, f_any="i", is_admin=True)
     verify_items(users, 2, [u5, u6])
-    users = get_users_for_index(session, False, f_any="i", expose_user_email=True, expose_user_name=True)
+    users = lib.get_users_for_index(session, False, f_any="i", expose_user_email=True, expose_user_name=True)
     verify_items(users, 2, [u5, u6])
-    users = get_users_for_index(session, False, f_any="i", expose_user_email=True)
+    users = lib.get_users_for_index(session, False, f_any="i", expose_user_email=True)
     verify_items(users, 1, [u5])
-    users = get_users_for_index(session, False, f_any="i", expose_user_name=True)
+    users = lib.get_users_for_index(session, False, f_any="i", expose_user_name=True)
     verify_items(users, 1, [u6])
 
     u1.deleted = True
-    users = get_users_for_index(session, True)
+    users = lib.get_users_for_index(session, True)
     verify_items(users, 1, [u1])
